Introduce ServerCall infrastructure
This commit adds the following classes (with extensive comments throughout the code), all enclosed in ServerApi namespace:
RequestConfig - encapsulates parameters needed to construct a network request;
Status - carried over from BaseJob
CallConfigBase and CallConfig<> - base classes to derive from in order to produce new types of server calls; supplied with several "reply loader" types that can be passed to CallConfig<> as a type parameter;
Call and PendingCall<> - the classes that encapsulate the server call itself: Call does heavy lifting of sending network requests and dealing with signals, while PendingCall<> is a template wrapper around it that allows to work with the response from the server in a type-safe way.
This commit also contains a proof-of-concept conversion of PasswordLogin, MediaThumbnail, JoinRoom, and Logout jobs to ServerApi and rewires Connection to use the new implementations.